refactor: test Controller 패키징
