+++ src/test/java/cart/controller/api/ProductApiControllerTest.java
@@ -1,4 +1,4 @@
-package cart.controller;
+package cart.controller.api;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
