+++ src/test/java/cart/controller/web/AdminControllerTest.java
@@ -1,4 +1,4 @@
-package cart.controller;
+package cart.controller.web;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.BDDMockito.given;
@@ -7,10 +7,10 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;
 
-import cart.controller.web.SettingController;
-import cart.domain.member.Member;
-import cart.dto.MembersResponse;
-import cart.service.MemberService;
+import cart.auth.LoginArgumentResolver;
+import cart.domain.product.Product;
+import cart.dto.ProductsResponse;
+import cart.service.ProductService;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -21,34 +21,36 @@ import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
 
-@WebMvcTest(SettingController.class)
-class SettingControllerTest {
+@WebMvcTest(AdminController.class)
+class AdminControllerTest {
 
     @Autowired
     private MockMvc mockMvc;
 
     @MockBean
-    private MemberService memberService;
+    private LoginArgumentResolver loginArgumentResolver;
 
-    @DisplayName("GET /settings")
+    @MockBean
+    private ProductService productService;
+
+    @DisplayName("GET /admin")
     @Test
-    void getSettings() throws Exception {
-        List<Member> members = List.of(
-                new Member((long) 1, "a@a.com", "abc1", "이오"),
-                new Member((long) 2, "b@b.com", "abc2", "애쉬")
-        );
-        given(memberService.findAll()).willReturn(members);
+    void getAdmin() throws Exception {
+        List<Product> products = List.of(
+                new Product((long) 1, "피자", 1000, "http://pizza"),
+                new Product((long) 2, "햄버거", 2000, "http://hamburger"));
+        given(productService.findAll()).willReturn(products);
 
-        MvcResult result = mockMvc.perform(get("/settings"))
+        MvcResult result = mockMvc.perform(get("/admin"))
                 .andExpect(status().isOk())
                 .andExpect(content().contentTypeCompatibleWith(MediaType.TEXT_HTML_VALUE))
-                .andExpect(view().name("settings"))
+                .andExpect(view().name("admin"))
                 .andReturn();
 
-        Object response = result.getModelAndView().getModel().get("members");
+        Object response = result.getModelAndView().getModel().get("products");
         assertThat(response)
-                .isInstanceOf(MembersResponse.class)
+                .isInstanceOf(ProductsResponse.class)
                 .usingRecursiveComparison()
-                .isEqualTo(MembersResponse.of(members));
+                .isEqualTo(ProductsResponse.of(products));
     }
 }
